Split up the package into separate files
