+++ response/successful.go
@@ -10,15 +10,6 @@ import (
 	"strconv"
 )
 
-type BadRequest struct {
-	DisplayText string
-}
-
-func (response BadRequest) WriteTo(client io.Writer) error {
-	writeStatusLine(client, 400, "Bad Request")
-	return nil
-}
-
 type DirectoryListing struct {
 	Files []os.FileInfo
 }
@@ -69,54 +60,4 @@ func contentTypeFromFileExtension(filename string) string {
 	}
 
 	return mime.TypeByExtension(extension)
-}
-
-type InternalServerError struct{}
-
-func (response InternalServerError) WriteTo(client io.Writer) {
-	writeStatusLine(client, 500, "Internal Server Error")
-}
-
-type NotFound struct {
-	Target string
-}
-
-func (response NotFound) WriteTo(client io.Writer) {
-	writeStatusLine(client, 404, "Not Found")
-	writeHeader(client, "Content-Type", "text/plain")
-
-	message := fmt.Sprintf("Not found: %s", response.Target)
-	writeHeader(client, "Content-Length", strconv.Itoa(len(message)))
-	writeEndOfMessageHeader(client)
-
-	writeBody(client, message)
-}
-
-type NotImplemented struct {
-	Method string
-}
-
-func (response NotImplemented) WriteTo(client io.Writer) error {
-	writeStatusLine(client, 501, "Not Implemented")
-	return nil
-}
-
-func writeStatusLine(client io.Writer, status int, reason string) {
-	fmt.Fprintf(client, "HTTP/1.1 %d %s\r\n", status, reason)
-}
-
-func writeHeader(client io.Writer, name string, value string) {
-	fmt.Fprintf(client, "%s: %s\r\n", name, value)
-}
-
-func writeEndOfMessageHeader(client io.Writer) {
-	fmt.Fprint(client, "\r\n")
-}
-
-func copyToBody(client io.Writer, bodyReader io.Reader) {
-	io.Copy(client, bodyReader)
-}
-
-func writeBody(client io.Writer, body string) {
-	fmt.Fprint(client, body)
 }
